Multiple bug fixes  - Handle Dx12 validation error related to multiple resource transitions happening on the same resource.  - Added API to better query for image/buffer views associated with an attachment.  - Address possible serialization issues by usage of size_t  - Reduced number of Descriptor sets allowed per SRG Pool in order to reduce descriptor waste.

+++ Gems/Atom/RPI/Code/Include/Atom/RPI.Edit/ResourcePool/ResourcePoolSourceData.h
@@ -37,7 +37,7 @@ namespace AZ
 
             ResourcePoolAssetType m_poolType = ResourcePoolAssetType::Unknown;
             AZStd::string m_poolName = "Unknown";
-            size_t m_budgetInBytes = 0;
+            int32_t m_budgetInBytes = 0;
 
             // Configuration for buffer pool
             RHI::HeapMemoryLevel m_heapMemoryLevel = RHI::HeapMemoryLevel::Device;
